requirements.txt
attrdict==2.0.1
gin_config==0.2.1
torch==1.1.0
scipy==1.3.1
numpy==1.17.2
opencv-python
opencv==3.4.2
.
